@@ -1,30 +1,19 @@
 package com.citadini.ourcity.filters;
 
-import java.io.IOException;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletResponse;
-
+import jakarta.servlet.*;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
+
+import java.io.IOException;
 
 @Component
 public class HeaderExposureFilter implements Filter {
 
 	@Override
-	public void init(FilterConfig filterConfig) throws ServletException {
-	}
-
-	@Override
-	public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-			throws IOException, ServletException {
-		HttpServletResponse res = (HttpServletResponse) response;
+	public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+		HttpServletResponse res = (HttpServletResponse) servletResponse;
 		res.addHeader("access-control-expose-headers", "location");
-		chain.doFilter(request, response);
+		filterChain.doFilter(servletRequest, res);
 	}
 
 	@Override
