@@ -1,48 +1,50 @@
 package com.citadini.ourcity.security;
 
-import java.io.IOException;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
 
-public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
+import java.io.IOException;
 
-	private JWTUtil jwtUtil;
+@Component
+public class JWTAuthorizationFilter extends OncePerRequestFilter {
+
+	private final JWTUtil jwtUtil;
 	
-	private UserDetailsService userDetailService;
+	private final UserDetailsService userDetailService;
 	
-	public JWTAuthorizationFilter(AuthenticationManager authenticationManager, JWTUtil jwtUtil, UserDetailsService userDetailService) {
-		super(authenticationManager);
+	public JWTAuthorizationFilter(JWTUtil jwtUtil, UserDetailsService userDetailService) {
+		super();
 		this.jwtUtil = jwtUtil;
 		this.userDetailService = userDetailService;
 	}
 
 	@Override
-	protected void doFilterInternal(HttpServletRequest request, 
-			HttpServletResponse response, 
-			FilterChain chain) throws IOException, ServletException {
+	protected void doFilterInternal(HttpServletRequest request,
+									HttpServletResponse response,
+									FilterChain filterChain) throws IOException, ServletException {
+
 		String header = request.getHeader("Authorization");
-		
+
 		if (header != null && header.startsWith("Bearer ")) {
 			UsernamePasswordAuthenticationToken auth = getAuthentication(header.substring(7));
 			if (auth != null) {
 				SecurityContextHolder.getContext().setAuthentication(auth);
 			}
 		}
-		chain.doFilter(request, response);
+		filterChain.doFilter(request, response);
 	}
 
 	private UsernamePasswordAuthenticationToken getAuthentication(String token) {
-		if (jwtUtil.tokenValido(token)) {
+		if (jwtUtil.tokenValidation(token)) {
 			String userName = jwtUtil.getUserName(token);
 			UserDetails user = userDetailService.loadUserByUsername(userName);
 			return new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
