@@ -1,14 +1,13 @@
 package com.citadini.ourcity.config;
 
-import java.text.ParseException;
-
+import com.citadini.ourcity.service.DBService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
-import com.citadini.ourcity.service.DBService;
+import java.text.ParseException;
 
 @Configuration
 @Profile("dev")
