@@ -1,14 +1,12 @@
 package com.citadini.ourcity.config;
 
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.regions.Regions;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 
 @Configuration
 public class S3Config {
@@ -20,14 +18,13 @@ public class S3Config {
 	private String awsKey;
 	
 	@Value("${s3.region}")
-	private String regio;
+	private String region;
 	
 	@Bean
 	public AmazonS3 s3client() {
 		BasicAWSCredentials awsCred = new BasicAWSCredentials(awsId, awsKey);
-		AmazonS3 s3client = AmazonS3ClientBuilder.standard().withRegion(Regions.fromName(regio))
+        return AmazonS3ClientBuilder.standard().withRegion(region)
 				.withCredentials(new AWSStaticCredentialsProvider(awsCred)).build();
-		return s3client;
 	}
 
 }
