@@ -1,43 +1,48 @@
 package com.citadini.ourcity.config;
 
-import java.util.Arrays;
-
+import com.citadini.ourcity.security.JWTAuthorizationFilter;
+import com.citadini.ourcity.security.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractAuthenticationFilterConfigurer;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.annotation.web.configurers.LogoutConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import com.citadini.ourcity.security.JWTAuthenticationFilter;
-import com.citadini.ourcity.security.JWTAuthorizationFilter;
-import com.citadini.ourcity.security.JWTUtil;
+import java.util.Arrays;
 
 @Configuration
 @EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+public class SecurityConfig {
 
 	@Autowired
-	private UserDetailsService userDetailService;
+	private UserDetailsService userDetailsService;
 	@Autowired
 	private Environment env;
 	@Autowired
 	private JWTUtil jwtUtil;
-	
+	@Autowired
+	private JWTAuthorizationFilter jwtAuthorizationFilter;
+
 	private static final String[] PUBLIC_MATCHERS = {
 			"/h2-console/**"
 	};
+
+	private static final String h2PublicRequest = "/h2-console/**";
 	
 	private static final String[] PUBLIC_MATCHERS_GET = {
 			"/categorias/**",
@@ -49,27 +54,29 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 			"/usuarios",
 			"/auth/forgot/**"
 	};
-	
-	@Override
-	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.userDetailsService(userDetailService).passwordEncoder(bCryptPasswordEncoder());
+
+	@Bean
+	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		http
+				.cors(AbstractHttpConfigurer::disable)
+				.csrf(AbstractHttpConfigurer::disable)
+				.sessionManagement(sm -> sm.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+				.authorizeHttpRequests(authorize -> authorize
+						.requestMatchers(h2PublicRequest).permitAll()
+						.requestMatchers(HttpMethod.GET, PUBLIC_MATCHERS_GET).permitAll()
+						.requestMatchers(HttpMethod.POST, PUBLIC_MATCHERS_POST).permitAll()
+						.anyRequest().authenticated()
+				)
+				.addFilterBefore(jwtAuthorizationFilter, UsernamePasswordAuthenticationFilter.class)
+				.formLogin(AbstractAuthenticationFilterConfigurer::permitAll)
+				.logout(LogoutConfigurer::permitAll);
+
+		return http.build();
 	}
-	
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		
-		if (Arrays.asList(env.getActiveProfiles()).contains("test"))
-			http.headers().frameOptions().disable();
-		
-		http.cors().and().csrf().disable();
-		http.authorizeRequests()
-			.antMatchers(PUBLIC_MATCHERS).permitAll()
-			.antMatchers(HttpMethod.POST, PUBLIC_MATCHERS_POST).permitAll()
-			.antMatchers(HttpMethod.GET, PUBLIC_MATCHERS_GET).permitAll()
-			.anyRequest().authenticated();
-		http.addFilter(new JWTAuthenticationFilter(authenticationManager(), jwtUtil));
-		http.addFilter(new JWTAuthorizationFilter(authenticationManager(), jwtUtil, userDetailService));
-		http.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+
+	@Bean
+	public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+		return configuration.getAuthenticationManager();
 	}
 	
 	@Bean
@@ -82,8 +89,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 	}
 	
 	@Bean
-	public BCryptPasswordEncoder bCryptPasswordEncoder() {
+	public BCryptPasswordEncoder passwordEncoder() {
 		return new BCryptPasswordEncoder();
 	}
-	
+
 }
