@@ -1,9 +1,8 @@
 package com.citadini.ourcity.dto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
 import java.io.Serializable;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotEmpty;
 
 public class EmailDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
