@@ -1,10 +1,9 @@
 package com.citadini.ourcity.dto;
 
-import java.io.Serializable;
-
-import javax.validation.constraints.NotEmpty;
-
 import org.hibernate.validator.constraints.Length;
+
+import jakarta.validation.constraints.NotEmpty;
+import java.io.Serializable;
 
 public class OcorrenciaDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
