@@ -1,14 +1,12 @@
 package com.citadini.ourcity.dto;
 
-import java.io.Serializable;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotEmpty;
-
-import org.hibernate.validator.constraints.Length;
-
 import com.citadini.ourcity.domain.Usuario;
 import com.citadini.ourcity.service.validation.UsuarioUpdate;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import org.hibernate.validator.constraints.Length;
+
+import java.io.Serializable;
 
 @UsuarioUpdate
 public class UsuarioDTO implements Serializable {
