@@ -1,9 +1,8 @@
 package com.citadini.ourcity.repositories;
 
+import com.citadini.ourcity.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.citadini.ourcity.domain.Categoria;
 
 @Repository
 public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
