@@ -1,10 +1,9 @@
 package com.citadini.ourcity.repositories;
 
+import com.citadini.ourcity.domain.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.citadini.ourcity.domain.Usuario;
 
 @Repository
 public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
