@@ -1,18 +1,13 @@
 package com.citadini.ourcity.resources;
 
-import java.util.List;
-
+import com.citadini.ourcity.domain.Categoria;
+import com.citadini.ourcity.service.CategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import com.citadini.ourcity.domain.Categoria;
-import com.citadini.ourcity.service.CategoriaService;
+import java.util.List;
 
 @RestController
 @RequestMapping(value = "/categorias")
