@@ -1,22 +1,5 @@
 package com.citadini.ourcity.resources;
 
-import java.net.URI;
-import java.util.List;
-
-import javax.validation.Valid;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import com.citadini.ourcity.domain.ApoioOcorrencia;
 import com.citadini.ourcity.domain.ComentarioOcorrencia;
 import com.citadini.ourcity.domain.Ocorrencia;
@@ -27,6 +10,16 @@ import com.citadini.ourcity.resources.utils.URL;
 import com.citadini.ourcity.service.ApoioOcorrenciaService;
 import com.citadini.ourcity.service.ComentarioOcorrenciaService;
 import com.citadini.ourcity.service.OcorrenciaService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import jakarta.validation.Valid;
+import java.net.URI;
+import java.util.List;
 
 @RestController
 @RequestMapping(value = "/ocorrencias")
