@@ -1,13 +1,13 @@
 package com.citadini.ourcity.security;
 
-import java.util.Date;
-
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
+import javax.crypto.SecretKey;
+import java.util.Date;
 
 @Component
 public class JWTUtil {
@@ -19,29 +19,34 @@ public class JWTUtil {
 	private Long expiration;
 	
 	public String generateToken(String username) {
+		SecretKey key = Keys.hmacShaKeyFor(secret.getBytes());
 		return Jwts.builder()
-				.setSubject(username)
-				.setExpiration(new Date(System.currentTimeMillis()+expiration))
-				.signWith(SignatureAlgorithm.HS512, secret.getBytes())
+				.subject(username)
+				.expiration(new Date(System.currentTimeMillis() + expiration))
+				.signWith(key) // New method
 				.compact();
 	}
 
-	public boolean tokenValido(String token) {
+	public boolean tokenValidation(String token) {
 		Claims claims = getClaims(token);
 		if (claims != null) {
 			String userName = claims.getSubject();
 			Date expirationDate = claims.getExpiration();
 			Date now = new Date(System.currentTimeMillis());
-			if (userName != null && expirationDate != null && now.before(expirationDate)) {
-				return true;
-			}
+            return userName != null && expirationDate != null && now.before(expirationDate);
 		}
 		return false;
 	}
 
 	private Claims getClaims(String token) {
 		try {
-			return Jwts.parser().setSigningKey(secret.getBytes()).parseClaimsJws(token).getBody();
+			SecretKey key = Keys.hmacShaKeyFor(secret.getBytes());
+
+			return Jwts.parser()
+					.verifyWith(key)  // Use the SecretKey object
+					.build()
+					.parseSignedClaims(token)
+					.getPayload();
 		} catch (Exception e) {
 			return null;
 		}
