@@ -1,17 +1,16 @@
 package com.citadini.ourcity.service;
 
-import java.util.List;
-import java.util.Optional;
-
+import com.citadini.ourcity.domain.Categoria;
+import com.citadini.ourcity.repositories.CategoriaRepository;
+import com.citadini.ourcity.service.exceptions.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 
-import com.citadini.ourcity.domain.Categoria;
-import com.citadini.ourcity.repositories.CategoriaRepository;
-import com.citadini.ourcity.service.exceptions.ObjectNotFoundException;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoriaService {
