@@ -1,16 +1,5 @@
 package com.citadini.ourcity.service;
 
-import java.util.Date;
-import java.util.Optional;
-
-import javax.validation.Valid;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort.Direction;
-import org.springframework.stereotype.Service;
-
 import com.citadini.ourcity.domain.ComentarioOcorrencia;
 import com.citadini.ourcity.domain.Ocorrencia;
 import com.citadini.ourcity.domain.Usuario;
@@ -19,6 +8,15 @@ import com.citadini.ourcity.repositories.ComentarioOcorrenciaRepository;
 import com.citadini.ourcity.security.UserSS;
 import com.citadini.ourcity.service.exceptions.AuthorizationException;
 import com.citadini.ourcity.service.exceptions.ObjectNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.stereotype.Service;
+
+import jakarta.validation.Valid;
+import java.util.Date;
+import java.util.Optional;
 
 @Service
 public class ComentarioOcorrenciaService {
