@@ -1,20 +1,5 @@
 package com.citadini.ourcity.service;
 
-import java.awt.image.BufferedImage;
-import java.net.URI;
-import java.util.Date;
-import java.util.Optional;
-
-import javax.validation.Valid;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort.Direction;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
 import com.citadini.ourcity.domain.Categoria;
 import com.citadini.ourcity.domain.Ocorrencia;
 import com.citadini.ourcity.domain.Status;
@@ -27,6 +12,19 @@ import com.citadini.ourcity.repositories.UsuarioRepository;
 import com.citadini.ourcity.security.UserSS;
 import com.citadini.ourcity.service.exceptions.AuthorizationException;
 import com.citadini.ourcity.service.exceptions.ObjectNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import jakarta.validation.Valid;
+import java.awt.image.BufferedImage;
+import java.net.URI;
+import java.util.Date;
+import java.util.Optional;
 
 @Service
 public class OcorrenciaService {
