@@ -1,14 +1,13 @@
 package com.citadini.ourcity.service;
 
+import com.citadini.ourcity.domain.Usuario;
+import com.citadini.ourcity.repositories.UsuarioRepository;
+import com.citadini.ourcity.security.UserSS;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import com.citadini.ourcity.domain.Usuario;
-import com.citadini.ourcity.repositories.UsuarioRepository;
-import com.citadini.ourcity.security.UserSS;
 
 @Service
 public class UserDetailServiceImpl implements UserDetailsService {
