@@ -1,12 +1,11 @@
 package com.citadini.ourcity.service.validation;
 
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
 
 @Constraint(validatedBy = OcorrenciaUpdateValidator.class)
 @Target({ ElementType.TYPE })
