@@ -1,17 +1,15 @@
 package com.citadini.ourcity.service.validation;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
 import com.citadini.ourcity.domain.Usuario;
 import com.citadini.ourcity.dto.UsuarioNewDTO;
 import com.citadini.ourcity.repositories.UsuarioRepository;
 import com.citadini.ourcity.resources.exceptions.FieldMessage;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+import java.util.ArrayList;
+import java.util.List;
 
 public class UsuarioInsertValidator implements ConstraintValidator<UsuarioInsert, UsuarioNewDTO> {
 	
